+++ app/mailers/inquiry_reply_mailer.rb
@@ -1,0 +1,11 @@
+class InquiryReplyMailer < ApplicationMailer
+  def inquiry_reply(inquiry_reply)
+    @inquiry_reply = inquiry_reply
+    mail(
+      from: ENV['SEND_MAIL'],  #送信元アドレス
+      to: contact.email,       #送信先アドレス
+      subject: 'お問い合わせを承りました',  #メールの件名
+      bcc: ENV['SEND_MAIL']    #BCC送信先アドレス
+    )
+  end
+end
