+++ spec/mailers/inquiry_reply_spec.rb
@@ -1,0 +1,5 @@
+require "rails_helper"
+
+RSpec.describe InquiryReplyMailer, type: :mailer do
+  pending "add some examples to (or delete) #{__FILE__}"
+end
