+++ spec/mailers/previews/inquiry_reply_preview.rb
@@ -1,0 +1,4 @@
+# Preview all emails at http://localhost:3000/rails/mailers/inquiry_reply
+class InquiryReplyPreview < ActionMailer::Preview
+
+end
